base: keep undo history private to dock
