+++ crates/ui/tests/base_compat.rs
@@ -385,7 +385,6 @@ fn element_ext_is_available_from_base_and_the_legacy_root() {
 #[test]
 fn legacy_history_path_reexports_the_base_type() {
     let _: gpui_component::history::History<u8> = gpui_base::History::new();
-    let _: gpui_component::history::UndoHistory<u8> = gpui_base::UndoHistory::new();
 }
 
 #[test]
