+++ website/base/index.md
@@ -28,4 +28,4 @@ The dependency points one way: `gpui-component` builds on `gpui-base`. Applicati
 
 Follow [Getting started](./getting-started.md), render selectable Markdown and HTML with [TextView](./text-view.md), learn how to add [window-level text selection](./text-selection.md) to custom renderers, then explore the [primitive catalog](./primitives/index.md). Each page includes Rust snippets and a live WASM example backed by the same example crate that can run natively.
 
-Three systems are larger than a primitive and have pages of their own. [Motion](./motion.md) provides typed transitions, springs, keyframes, presence, and sequencing. [Virtual List](./virtual-list.md) renders lists of any length by drawing only what is on screen, with per-item sizes rather than a uniform row height. [Dock](./dock.md) is a full workspace shell — nested splits, tab groups, a tiles canvas and edge docks — whose layout is pure data you can build and serialize without a window, and whose every pixel comes from renderer traits you implement. [History and Undo History](./history.md) are smaller but deliberately distinct: `History` is a root/current/back/forward navigation trail, while `UndoHistory` records grouped undo and redo transactions.
+Three systems are larger than a primitive and have pages of their own. [Motion](./motion.md) provides typed transitions, springs, keyframes, presence, and sequencing. [Virtual List](./virtual-list.md) renders lists of any length by drawing only what is on screen, with per-item sizes rather than a uniform row height. [Dock](./dock.md) is a full workspace shell — nested splits, tab groups, a tiles canvas and edge docks — whose layout is pure data you can build and serialize without a window, and whose every pixel comes from renderer traits you implement. [History](./history.md) is a root/current/back/forward navigation trail.
